pyproject.toml
[project]
name = "chatapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.6",
    "python-dotenv>=1.2.2",
]
